Cast to float

+++ app/utilities/apl_legacy_handicap_system.py
@@ -107,9 +107,11 @@ class APLLegacyHandicapSystem(WorldHandicapSystem):
             score_diffs_avg = np.mean(record_sorted[0:4])
         else:
             score_diffs_avg = np.mean(record_sorted[0:5])
-        return min(
-            np.floor((0.96 * score_diffs_avg) * 10.0) / 10.0,
-            self.maximum_handicap_index,
+        return float(
+            min(
+                np.floor((0.96 * score_diffs_avg) * 10.0) / 10.0,
+                self.maximum_handicap_index,
+            )
         )  # truncate to nearest tenth
 
     def compute_match_hole_result(
